Change the errors for an edge case test
We still produce the same parse tree, but we don't arrive at the same
part of the parse result that would have emitted an error about a module
definition in a method body.

But also, that error doesn't really make sense here? This module is
clearly not inside a method body.

So I think we're fine to just update the test and move on.