Also account for inheritance
The `!= untyped` guard is also there because before checking for
`isClassOrModuleFinal`, some code in `Module#===` would also hit this
case.

Technically speaking, `Symbols::untyped` is not `final!`, so the
condition is redundant, but in the process of debugging I learned that
`attachedClass()` on `Symbols::untyped()` is defined to return itself,
so I figured I may as well make that intention explicit here.