Add generic "error_node" builder
I've chosen to implement this as an (unresolved) constant reference. I
think that making a fake constant node will be easier than having to
update all the places in the codebase that might want to exhaustively
handle all pieces of syntax.

I've chosen to make it a constant node because it's allowed in more
places (versus say, a variable name or method name).

Another option was to make it an empty `begin; end` node, but that would
become an `EmptyTree` after desugar, and I think that this is more
useful in case we want to detect this somewhere later in the pipeline.